Clean up file structure and modules
