+++ src/sse/authenticated-sse-transport.ts
